% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire spreadsheet}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{distance of the fire}
  \item{DAMAGE}{total damage the fire did}
}
}
\usage{
fire
}
\description{
Fire data
}
\keyword{datasets}
