% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{myboot}
\alias{myboot}
\title{Myboot1}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{Number of columns}

\item{x}{Sample data}

\item{fun}{function we are analyzing}

\item{alpha}{confidence interval alpha}

\item{...}{Extra implementation}
}
\value{
Plot of data related to fun
}
\description{
Myboot1
}
\examples{
myboot(x = c(10, 20, 30))
}
