% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{myboot2}
\alias{myboot2}
\title{Myboot2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of columns}

\item{x}{How many samples are taken}

\item{fun}{Function that we are implementing}

\item{alpha}{Range of error}

\item{cx}{Location of text compared to density}

\item{...}{Extra implmentation}
}
\value{
Plot of data related to fun
}
\description{
Myboot2
}
\examples{
myboot2(x = 20)
}
