% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{myclt}
\alias{myclt}
\title{Random number histogram generator}
\usage{
myclt(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{Number of rows in data matrix}

\item{iter}{Number of columns in data matrix}

\item{a}{Lower bound}

\item{b}{Upper bound}
}
\value{
Histogram of data matrix
}
\description{
Random number histogram generator
}
\examples{
myclt(n=50,iter=10000,a=5,b=10)
}
