% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mycltu.R
\name{mycltu}
\alias{mycltu}
\title{Central limit function for the uniform distribution}
\usage{
mycltu(n, iter, a = 0, b = 10)
}
\arguments{
\item{n}{number of samples}

\item{iter}{number of iterations}

\item{a}{lower limit of the uniform distribution}

\item{b}{upper limit of the uniform distirbution}
}
\value{
returns a histogram of the sample distribution, and its theorical density curve and the density curve made from the data
}
\description{
Central limit function for the uniform distribution
}
\examples{
mycltu(n=4,iter=1000)
}
