% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfun.R
\name{myfun}
\alias{myfun}
\title{Square a vector}
\usage{
myfun(x)
}
\arguments{
\item{x}{A vector of quantitative values}
}
\value{
A vector of squared values
}
\description{
Square a vector
}
\examples{
myfun(x = 1:4)
}
