% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Math.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum parameter finder}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{function we are analyzing}

\item{x}{parameter that we provide values for}

\item{param}{single parameter that we are looking to find the max value for}

\item{...}{extra}
}
\value{
Graph that shows where max value is on the graph
}
\description{
Maximum parameter finder
}
\examples{
mymaxlik(lfun = (ex = function(x, param){ mean(x)/param}), x=c(3,4,5,2,3,4,5,5,5,1), param = seq(0,1,length=1000))
}
