% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Number of Tickets function}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Number of seats on the plane}

\item{gamma}{percentage of overflow for this flight based on # of seats}

\item{p}{percentage of passenger not showing up based on # of seats}
}
\value{
list of values
}
\description{
Number of Tickets function
}
\examples{
ntickets(200, 0.02, 0.95)
}
